+++ internal/http/interceptors/cors/cors.go
@@ -21,7 +21,7 @@ package cors
 import (
 	"net/http"
 
-	"github.com/cs3org/reva/pkg/rhttp/global"
+	"github.com/cs3org/reva/pkg/rhttp"
 	"github.com/mitchellh/mapstructure"
 	"github.com/rs/cors"
 )
@@ -31,7 +31,7 @@ const (
 )
 
 func init() {
-	global.RegisterMiddleware("cors", New)
+	rhttp.RegisterMiddleware("cors", New)
 }
 
 type config struct {
@@ -47,7 +47,7 @@ type config struct {
 }
 
 // New creates a new CORS middleware.
-func New(m map[string]interface{}) (global.Middleware, int, error) {
+func New(m map[string]interface{}) (rhttp.Middleware, int, error) {
 	conf := &config{}
 	if err := mapstructure.Decode(m, conf); err != nil {
 		return nil, 0, err
