+++ internal/http/services/metrics/metrics.go
@@ -29,13 +29,13 @@ import (
 	"github.com/cs3org/reva/pkg/logger"
 	"github.com/cs3org/reva/pkg/metrics"
 	"github.com/cs3org/reva/pkg/metrics/config"
-	"github.com/cs3org/reva/pkg/rhttp/global"
+	"github.com/cs3org/reva/pkg/rhttp"
 	"github.com/cs3org/reva/pkg/rhttp/mux"
 	"github.com/cs3org/reva/pkg/utils/cfg"
 )
 
 func init() {
-	global.Register(serviceName, New)
+	rhttp.Register(serviceName, New)
 }
 
 const (
@@ -64,7 +64,7 @@ func (s *svc) Register(r mux.Router) {
 }
 
 // New returns a new metrics service.
-func New(ctx context.Context, m map[string]interface{}) (global.Service, error) {
+func New(ctx context.Context, m map[string]interface{}) (rhttp.Service, error) {
 	var c config.Config
 	if err := cfg.Decode(m, &c); err != nil {
 		return nil, err
