+++ internal/http/services/reverseproxy/reverseproxy.go
@@ -27,7 +27,7 @@ import (
 	"os"
 
 	ctxpkg "github.com/cs3org/reva/pkg/ctx"
-	"github.com/cs3org/reva/pkg/rhttp/global"
+	"github.com/cs3org/reva/pkg/rhttp"
 	"github.com/cs3org/reva/pkg/rhttp/mux"
 	"github.com/cs3org/reva/pkg/utils/cfg"
 )
@@ -35,7 +35,7 @@ import (
 const name = "reverseproxy"
 
 func init() {
-	global.Register(name, New)
+	rhttp.Register(name, New)
 }
 
 type proxyRule struct {
@@ -58,7 +58,7 @@ type svc struct {
 }
 
 // New returns an instance of the reverse proxy service.
-func New(ctx context.Context, m map[string]interface{}) (global.Service, error) {
+func New(ctx context.Context, m map[string]interface{}) (rhttp.Service, error) {
 	var c config
 	if err := cfg.Decode(m, &c); err != nil {
 		return nil, err
