+++ internal/http/services/siteacc/siteacc.go
@@ -23,7 +23,7 @@ import (
 	"net/http"
 
 	"github.com/cs3org/reva/pkg/appctx"
-	"github.com/cs3org/reva/pkg/rhttp/global"
+	"github.com/cs3org/reva/pkg/rhttp"
 	"github.com/cs3org/reva/pkg/rhttp/mux"
 	"github.com/cs3org/reva/pkg/siteacc"
 	"github.com/cs3org/reva/pkg/siteacc/config"
@@ -33,7 +33,7 @@ import (
 )
 
 func init() {
-	global.Register(serviceName, New)
+	rhttp.Register(serviceName, New)
 }
 
 type svc struct {
@@ -86,7 +86,7 @@ func applyDefaultConfig(conf *config.Configuration) {
 }
 
 // New returns a new Site Accounts service.
-func New(ctx context.Context, m map[string]interface{}) (global.Service, error) {
+func New(ctx context.Context, m map[string]interface{}) (rhttp.Service, error) {
 	var c config.Configuration
 	if err := cfg.Decode(m, &c); err != nil {
 		return nil, err
