+++ internal/http/services/sysinfo/sysinfo.go
@@ -23,14 +23,14 @@ import (
 	"net/http"
 
 	"github.com/cs3org/reva/pkg/appctx"
-	"github.com/cs3org/reva/pkg/rhttp/global"
+	"github.com/cs3org/reva/pkg/rhttp"
 	"github.com/cs3org/reva/pkg/rhttp/mux"
 	"github.com/cs3org/reva/pkg/sysinfo"
 	"github.com/cs3org/reva/pkg/utils/cfg"
 )
 
 func init() {
-	global.Register(serviceName, New)
+	rhttp.Register(serviceName, New)
 }
 
 type config struct {
@@ -88,7 +88,7 @@ func (s *svc) getJSONData() string {
 }
 
 // New returns a new SysInfo service.
-func New(ctx context.Context, m map[string]interface{}) (global.Service, error) {
+func New(ctx context.Context, m map[string]interface{}) (rhttp.Service, error) {
 	var c config
 	if err := cfg.Decode(m, &c); err != nil {
 		return nil, err
