--- pkg/rhttp/global/global.go
@@ -1,59 +1,0 @@
-// Copyright 2018-2023 CERN
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-//
-// In applying this license, CERN does not waive the privileges and immunities
-// granted to it by virtue of its status as an Intergovernmental Organization
-// or submit itself to any jurisdiction.
-
-package global
-
-import (
-	"context"
-	"io"
-	"net/http"
-
-	"github.com/cs3org/reva/pkg/rhttp/mux"
-)
-
-// NewMiddlewares contains all the registered new middleware functions.
-var NewMiddlewares = map[string]NewMiddleware{}
-
-// NewMiddleware is the function that HTTP middlewares need to register at init time.
-type NewMiddleware func(conf map[string]interface{}) (Middleware, int, error)
-
-// RegisterMiddleware registers a new HTTP middleware and its new function.
-func RegisterMiddleware(name string, n NewMiddleware) {
-	NewMiddlewares[name] = n
-}
-
-// Middleware is a middleware http handler.
-type Middleware func(h http.Handler) http.Handler
-
-// Services is a map of service name and its new function.
-var Services = map[string]NewService{}
-
-// Register registers a new HTTP services with name and new function.
-func Register(name string, newFunc NewService) {
-	Services[name] = newFunc
-}
-
-// NewService is the function that HTTP services need to register at init time.
-type NewService func(context.Context, map[string]interface{}) (Service, error)
-
-// Service represents a HTTP service.
-type Service interface {
-	Name() string
-	Register(r mux.Router)
-	io.Closer
-}
